+++ R/enrichmap.R
@@ -43,7 +43,7 @@
 #'   \code{x[[statistic_column]]}, respectively. Default is "#3366ff" (blue) and
 #'   "darkred".
 #' @param heatmap_color_fun function; used to create the legend for the heatmap
-#'   bubble fill. See \code{\link{enrichmap_color_functions}} for details.
+#'   bubble fill. See \code{\link{enrichmapColorFunctions}} for details.
 #' @param scale_by character; whether to scale the bubbles such that the term
 #'   with the largest \eqn{-log_{10}} adjusted p-value in each row
 #'   (\code{scale_by="row"}), column (\code{scale_by="column"}), or overall
@@ -135,14 +135,13 @@ enrichmap <- function(x,
                       contrast_column = "Contrast",
                       padj_column = "FDR",
                       padj_legend_title = padj_column,
-                      padj_aggregate_fun = function(padj) {
-                          median(-log10(padj), na.rm = TRUE)
-                      },
+                      padj_aggregate_fun = function(padj)
+                          median(-log10(padj), na.rm = TRUE),
                       padj_cutoff = 0.05,
                       plot_sig_only = TRUE,
                       padj_fill = "grey",
                       colors = c("#3366ff", "darkred"),
-                      heatmap_color_fun = camera_color_fun,
+                      heatmap_color_fun = cameraColorFun,
                       scale_by = c("row", "column", "max"),
                       cell_size = unit(14, "points"),
                       filename,
@@ -152,8 +151,7 @@ enrichmap <- function(x,
                       heatmap_args = list(),
                       padj_args = list(),
                       save_args = list(),
-                      draw_args = list())
-{
+                      draw_args = list()) {
     scale_by <- match.arg(scale_by,
                           choices = c("row", "column", "max"))