+++ R/readGMT.R
@@ -14,7 +14,7 @@
 #'
 #' @returns A named list of character vectors.
 #'
-#' @export gmt_to_list
+#' @export readGMT
 #'
 #' @note Similar to \code{fgsea::gmtPathways}.
 #'
@@ -22,13 +22,13 @@
 #' path <- system.file("extdata", "c5.go.v2023.2.Hs.symbols.gmt.gz",
 #'                     package = "TMSig")
 #'
-#' x <- gmt_to_list(path)
+#' x <- readGMT(path)
 #'
 #' head(names(x)) # First 6 gene set names
 #'
 #' x[1] # first set
 
-gmt_to_list <- function(path, check = TRUE) {
+readGMT <- function(path, check = TRUE) {
     if (check & !grepl("\\.gmt(\\.[^\\.]+)?$", path))
         stop("`path` is not a path to a GMT file.")